==> notebooks/4 - using data APIs.ipynb <==
import requests

### `Flask API` (NOT NEEDED)

url = "http://localhost:5000/api/v1/resources/orders/tx"

response = requests.request("GET", url)
print("STATUS: {}".format(response))

content = response.json()
content

### `FastAPI`

url = "http://localhost:8000/api/v1/resources/locations/tx"

response = requests.request("GET", url)
print("STATUS: {}".format(response))

content = response.json()
content

url = "http://localhost:8000/api/v1/resources/orders/tx"

response = requests.request("GET", url)
print("STATUS: {}".format(response))

content = response.json()
content
