==> notebooks/5 - MySQL data.ipynb <==
import os

import numpy as np
import pandas as pd

from getpass import getpass
from mysql.connector import connect, Error

data_dir = os.path.join("..", "Datasets", "db")
customer_t = pd.read_csv(os.path.join(data_dir, "Customers.csv"))
location_t = pd.read_csv(os.path.join(data_dir, "Locations.csv"))
order_detail_t = pd.read_csv(os.path.join(data_dir, "Order_Details.csv"))
order_t = pd.read_csv(os.path.join(data_dir, "Orders.csv"))
product_t = pd.read_csv(os.path.join(data_dir, "Products.csv"))

product_t[product_t["product_name"].duplicated()]

product_t.nunique()

order_t

product_t['product_name'].nunique()

customer_t

## Data Preparation
# customer_t = customer_t.rename(columns={"customer_id": "customer_code"})
# customer_t["customer_id"] = customer_t.index
customer_t = customer_t[["customer_id", "customer_name", "segment", "email", "password", "username"]]

product_t = product_t[["product_id", "product_name", "product_code", "category", "sub_category", "product_price"]]
product_t["product_id"] = product_t["product_id"].astype("str")

location_t = location_t[["location_id", "country", "city", "state", "postal_code", "region"]]
location_t["location_id"] = location_t["location_id"].astype("str")
location_t["postal_code"] = location_t["postal_code"].astype("str")

# order_t = order_t.rename(columns={"order_id": "order_code"})
# order_t['order_id'] = order_t.index
order_t = order_t[["order_id", "order_date", "ship_date", "ship_mode", "customer_id"]]

order_detail_t = order_detail_t[["order_detail_id", "order_id", "product_id", "location_id", "sales", "quantity", 
                                     "discount", "profit"]]
order_detail_t["order_detail_id"] = order_detail_t["order_detail_id"].astype("str")
order_detail_t["product_id"] = order_detail_t["product_id"].astype("str")
order_detail_t["location_id"] = order_detail_t["location_id"].astype("str")
# customer_t["customer_id"] = customer_t["customer_id"].astype("str")
# order_t["order_id"] = order_t["order_id"].astype("str")
order_detail_t["sales"] = order_detail_t["sales"].astype("str")
order_detail_t["quantity"] = order_detail_t["quantity"].astype("str")
order_detail_t["discount"] = order_detail_t["discount"].astype("str")
order_detail_t["profit"] = order_detail_t["profit"].astype("str")

### `Connecting to MySQL`

username = "root"
password = "123m"

try:
    conn = connect(
        host="localhost",
        user=username,
        password=password,
        port="3306"
    ) 
    print(conn)
except Error as e:
    print(e)

curr = conn.cursor()

### `Setup DB`

db_name = "sales_db"

sql_create_db = "CREATE DATABASE IF NOT EXISTS {};".format(db_name)

curr.execute(sql_create_db)

# sql_use_db = "USE sales_db;"

# curr.execute(sql_use_db)

# RE-CONNECTION: specific to database only

try:
    conn = connect(
        host="localhost",
        user=username,
        password=password,
        database=db_name
    ) 
    print(conn)
except Error as e:
    print(e)

curr = conn.cursor()

### `Table Creation`

order_t

# customers table
sql_create_table_customers = """
    CREATE TABLE IF NOT EXISTS Customers(
        row_id INT AUTO_INCREMENT,
        customer_id varchar(75) PRIMARY KEY,
        customer_name varchar(250) NOT NULL,
        segment varchar(100) NOT NULL,
        email varchar(250),
        password varchar(250),
        username varchar(200),
        INDEX(row_id)
    );
"""

# products table
sql_create_table_products = """
    CREATE TABLE IF NOT EXISTS Products(
        row_id INT unsigned PRIMARY KEY,
        product_name varchar(300) NOT NULL,
        product_code varchar(300) NOT NULL,
        category varchar(100),
        sub_category varchar(100),
        product_price float
    );
"""

# locations table
sql_create_table_locations = """
    CREATE TABLE IF NOT EXISTS Locations(
        row_id INT unsigned PRIMARY KEY,
        country varchar(150) NOT NULL,
        city varchar(150) NOT NULL,
        state varchar(150),
        postal_code varchar(50),
        region varchar(50)
    );
"""

# orders table
sql_create_table_orders = """
    CREATE TABLE IF NOT EXISTS Orders(
        row_id int AUTO_INCREMENT,
        order_id varchar(75) PRIMARY KEY,
        order_date date NOT NULL,
        ship_date date,
        ship_mode varchar(50),
        customer varchar(75),
        INDEX(row_id), 
        FOREIGN KEY (customer) REFERENCES Customers(customer_id)
    );
"""

# order details table
sql_create_table_order_details = """
    CREATE TABLE IF NOT EXISTS OrderDetails(
        row_id int unsigned PRIMARY KEY,
        order_id varchar(75) NOT NULL,
        product_bought int unsigned NOT NULL,
        location_bought int unsigned NOT NULL,
        sales float NOT NULL,
        quantity int NOT NULL,
        discount float,
        profit float,
        FOREIGN KEY (order_id) REFERENCES Orders(order_id),
        FOREIGN KEY (product_bought) REFERENCES Products(row_id),
        FOREIGN KEY (location_bought) REFERENCES Locations(row_id)
    );
"""

curr.execute(sql_create_table_customers)
curr.execute(sql_create_table_products)
curr.execute(sql_create_table_locations)
curr.execute(sql_create_table_orders)
curr.execute(sql_create_table_order_details)

### `Data Insertion`

# customers insertion
sql_insert_customers = """
    INSERT INTO Customers
        (customer_id, customer_name, segment, email, password, username)
        VALUES (%s, %s, %s, %s, %s, %s)
"""
curr.executemany(sql_insert_customers, customer_t.values.tolist())


# products insertion
sql_insert_products = """
    INSERT INTO Products
        (row_id, product_name, product_code, category, sub_category, product_price)
        VALUES (%s, %s, %s, %s, %s, %s)
        
"""
curr.executemany(sql_insert_products, product_t.values.tolist())


# locations insertion
sql_insert_locations = """
    INSERT INTO Locations
        (row_id, country, city, state, postal_code, region)
        VALUES (%s, %s, %s, %s, %s, %s)
        
"""
curr.executemany(sql_insert_locations, location_t.values.tolist())

# orders insertion
sql_insert_orders = """
    INSERT INTO Orders
        (order_id, order_date, ship_date, ship_mode, customer)
        VALUES (%s, %s, %s, %s, %s)
"""
curr.executemany(sql_insert_orders, order_t.values.tolist())

# order details insertion
sql_insert_order_details = """
    INSERT INTO OrderDetails
        (row_id, order_id, product_bought, location_bought, sales, quantity, discount, profit)
        VALUES (%s, %s, %s, %s, %s, %s, %s, %s)
        
"""
curr.executemany(sql_insert_order_details, order_detail_t.values.tolist())

conn.commit()
