==> notebooks/7 - eCommerce API - dojo shopping data.ipynb <==
`about`:  PUBLIC data and features of related site/brand "dojo". 

query for all information about categories, products, etc… as on the official website. 

import requests
import unicodedata
import re
import random
import pandas as pd

# helper library
from ast import literal_eval
from bs4 import BeautifulSoup

## 0- Product Categories 

import requests

url = "https://asos2.p.rapidapi.com/categories/list"

querystring = {"country":"US","lang":"en-US"}

headers = {
    "X-RapidAPI-Key": "d52cef56e7msh212a5f9ac2d768dp169c65jsna0a06fdb586f",
    "X-RapidAPI-Host": "asos2.p.rapidapi.com"
}

response = requests.request("GET", url, headers=headers, params=querystring)


content_pc = response.json()
content_pc.keys()

men_nav['children'][4]['children'][3]
men_nav['children'][4]['children'][6]
men_nav['children'][4]['children'][7]
men_nav['children'][4]['children'][10]
men_nav['children'][4]['children'][11]
men_nav['children'][4]['children'][13]

main_categories = [
    men_nav['children'][4]['children'][3],
    women_nav['children'][4]['children'][3],
]

# IMPORTANT
category_ids = []
categories = []
sub_categories = []


for mc in main_categories:
    main_category_name = mc['content']['title']
    print(main_category_name)
    
    mc_products = mc['children'][1]['children'][3:]
    for mc_product in mc_products:
        mc_subcategory = mc_product['content']['title']
        mc_cate_id = mc_product['link']['categoryId']
        print(" - {}".format(mc_subcategory), end=" / ")
        print("{}".format(mc_cate_id))
        category_ids.append(mc_cate_id)
        categories.append("clothing")
        sub_categories.append(mc_subcategory)
    print()

# men_nav['children'][4]['children'][3]['content']['title']

# men_nav['children'][4]['children'][3]['children'][1]['children'][3]['content']['title']
men_nav['children'][4]['children'][3]['children'][1]['children'][4]


print()

men_nav['children'][4]['children'][3]['children'][1]['children'][2]

# navigation
navigation = content_pc['navigation']

men_nav = navigation[0]
print("NAV 1 - {}".format(men_nav['content']['title']))
print("({} Items)".format(len(men_nav['children'])))
for item in men_nav['children']:
    print(" - {}".format(item['content']['title']))
print()

women_nav = navigation[1]
print("NAV 2 - {}".format(women_nav['content']['title']))
print("({} Items)".format(len(women_nav['children'])))
for item in women_nav['children']:
    print(" - {}".format(item['content']['title']))
print()

# brands
brands = content_pc['brands']
category_ids = []

for brand_i in brands:
    main_category_name = brand_i['content']['title']
    print("Category Name: {}".format(main_category_name))
    
    category_brands = brand_i['children']
    print("Number of Items: {}".format(len(category_brands)))
    for category_brand_i in category_brands:
        category_brand = category_brand_i['content']['title']
        category_brand_id = category_brand_i['link']['categoryId']
        category_ids.append((category_brand_id, category_brand, main_category_name))
    print()
    
# IMPORTANT
category_ids[:5]

# footer
print("FOOTER")
for f_item in content_pc['footer']:
    print(" - {}".format(f_item['content']['title']))

## 1- Products list

#### 1.1 Shoes, Boots & Sneakers

url = "https://asos2.p.rapidapi.com/products/v2/list"

querystring = {"store":"US",
               "offset":"0",
               "categoryId":"4209",  # FACTOR
               "limit":"48",
               "country":"US",
               "sort":"freshness",
               "currency":"USD",
               "sizeSchema":"US",
               "lang":"en-US"
              }

headers = {
    "X-RapidAPI-Key": "d52cef56e7msh212a5f9ac2d768dp169c65jsna0a06fdb586f",
    "X-RapidAPI-Host": "asos2.p.rapidapi.com"
}

response = requests.request("GET", url, headers=headers, params=querystring)
content = response.json()

# Get content of interest
products_list = content['products']
cloth_facets = content['facets']
campaigns_related = content['campaigns']  # empty

# meta data
meta_diagnostics = content['diagnostics']
meta_item_count = content['itemCount']   # how is item counted calculated?
meta_query_id = content['queryId']

# 
print("Number of products: {}".format(len(products_list)))

print('Product Category: {}'.format(content['categoryName']))

print('Product Data (E.g. each item)')

eg_product_item = products_list[0]
print("- ID: {}".format(eg_product_item['id']))
print("- Name: {}".format(eg_product_item['name']))
print("- Product Code: {}".format(eg_product_item['productCode']))
print("- Product Type: {}".format(eg_product_item['productType']))
print("- Brand Name: {}".format(eg_product_item['brandName']))
print("- Product Color: {}".format(eg_product_item['colour']))    
print("- Current Price: {}".format(eg_product_item['price']['current']['value']))
print('- Currency of Price: {}'.format(eg_product_item['price']['currency']))
print()

print("- Url of Product: \n'{}'".format(eg_product_item['url']))
print("- Image Url of Product: \n'{}'".format(eg_product_item['imageUrl']))
print()

print("- Product has variant colours? {}".format(eg_product_item['hasVariantColours']))
print("- Product has different prices? {}".format(eg_product_item['hasMultiplePrices']))
print("- Product Price marked down? {}".format(eg_product_item['price']['isMarkedDown']))
print("- Product Price same as outlet price? {}".format(eg_product_item['price']['isOutletPrice']))

print('Product Facets (related-features)')
for i in cloth_facets:
    print("- {}".format(i['id']))

len(category_ids)

# IMPORTANT
# n_items = 100
# category_list_ids_sample = list(map(lambda x: str(x[0]), random.sample(category_ids, n_items)))
# category_list_ids_sample[:4]
category_list_ids_sample = category_ids

# IMPORTANT
product_data = []
product_cols = ["product_code", "product_name", "product_type", "product_subtype", "product_price"]

for cate_idx in range(len(category_list_ids_sample)):
    cate_id = category_ids[cate_idx]
    sub_category = sub_categories[cate_idx]
    category = categories[cate_idx]
    
    querystring["categoryId"] = cate_id
    response_i = requests.request("GET", url, headers=headers, params=querystring)
    content_i = response_i.json()
    
    products_list = content_i['products']
    for product_i in products_list:
        product_name = product_i['name']
        # product_type = "Clothes"
        # product_subtype = content_i['categoryName']
        product_price = product_i['price']['current']['value']
        product_code = product_type[:3].upper()+ "-" + product_subtype[:2].upper() + "-"+ \
                                    str(product_i['productCode'])[:8]
        product_data.append((product_code, product_name, category, sub_category, product_price))
    

product_new_df = pd.DataFrame(product_data, columns=product_cols)

product_new_df["product_subtype"].unique()

product_new_df.to_csv("../Datasets/shopping_product_data.csv", index=False)

# IMPORTANT (test)
product_name = products_list[0]['name']
print(product_name)

product_type = "Clothes"
print(product_type)

product_subtype = content['categoryName']
print(product_subtype)

product_price = products_list[0]['price']['current']['value']
print(product_price)

product_code = product_type[:3].upper()+ "-" + product_subtype[:2].upper() + "-"+ \
                                    str(products_list[0]['productCode'])[:8]
print(product_code)

## 2- Product Details

enter product id to get product specific details

url = "https://asos2.p.rapidapi.com/products/v3/detail"

querystring = {"id":"9851613", 
               "lang":"en-US",
               "store":"US",
               "sizeSchema":"US",
               "currency":"USD"}

headers = {
    "X-RapidAPI-Key": "d52cef56e7msh212a5f9ac2d768dp169c65jsna0a06fdb586f",
    "X-RapidAPI-Host": "asos2.p.rapidapi.com"
}

response = requests.request("GET", url, headers=headers, params=querystring)
content_pd = response.json()

content_pd.keys()

# text cleaning functions 
# 1
CLEANR = re.compile('<.*?>') 
def cleanhtml(raw_html):
    cleantext = re.sub(CLEANR, ' ', raw_html)
    return cleantext

# 2
# case: remove non-breaking space in Latin1
def normalize_unicode(text):
    text = unicodedata.normalize("NFKD", text)
    return text

# 3
multiple_space_re = re.compile("\s{2,}")
def remove_multiple_whitespace(text):
    return multiple_space_re.sub(repl=" ", string=text)

def remove_lead_trail_whitespace(text):
    return text.strip()

# all data needed
print("product ID: {}".format(content_pd['id']))
print("product Name: {}".format(content_pd['name']))
remove_lead_trail_whitespace(remove_multiple_whitespace(normalize_unicode(cleanhtml(content_pd['description']))))

print("Gender: {}".format(content_pd['gender']))
print("product Code: {}".format(content_pd['productCode']))
print("product ??: {}".format(content_pd['pdpLayout']))
print("Brand: {}".format(content_pd['brand']['name']))
content_pd['isInStock']
content_pd['countryOfManufacture']
content_pd['price']
content_pd['rating']
content_pd['baseUrl']

# note: needs to be unpacked in multiple rows
content_pd['variants']
content_pd['media']
content_pd['info']
print("product Type: {}".format(content_pd['productType']['name']))
content_pd['webCategories']

content_pd['hasVariantsWithProp65Risk']
content_pd['isNoSize']
content_pd['isOneSize']

# # IMPORTANT
# product_name = content_pd['name']
# print(product_name)

# product_type = content_pd['price']['current']['value']
# print(product_type)

# sub_type = content_pd['productType']['name']
# print(sub_type)

## 3- dojo countries + categories

#### 3.1 categories

import requests

url = "https://asos2.p.rapidapi.com/categories/list"

querystring = {"country":"US","lang":"en-US"}

headers = {
    "X-RapidAPI-Key": "d52cef56e7msh212a5f9ac2d768dp169c65jsna0a06fdb586f",
    "X-RapidAPI-Host": "asos2.p.rapidapi.com"
}

response = requests.request("GET", url, headers=headers, params=querystring)
content = response.json()

len(content['brands'])

#### 3.2 countries

url = "https://asos2.p.rapidapi.com/countries/list"

querystring = {"lang":"en-US"}

headers = {
    "X-RapidAPI-Key": "d52cef56e7msh212a5f9ac2d768dp169c65jsna0a06fdb586f",
    "X-RapidAPI-Host": "asos2.p.rapidapi.com"
}

response = requests.request("GET", url, headers=headers, params=querystring)
content = response.json()

print('number of countries: {}'.format(len(content)))

API link:
- https://rapidapi.com/apidojo/api/asos2/
