==> notebooks/9 - MongoDB Data.ipynb <==
import numpy as np
import pandas as pd 

import re
from pymongo import MongoClient

import math
from datetime import date

def get_database(db_name):
    CONNECTION_STRING = "mongodb://modaj:123m@localhost:27017/{}".format(db_name)
    client = MongoClient(CONNECTION_STRING)
    return client[db_name] 

symb_re3 = re.compile(r"""[^a-zA-Z0-9\s]+""")

symb_re2 = re.compile(r"""[\W_]+""")

ini_string = "59 year abc555jw:, .’’’’@! eiw"
 
# printing initial string
print ("initial string : ", ini_string)

result = symb_re3.sub(repl="", string=ini_string)

# printing final string
print ("final string", result)

ini_string = "abc555jw:, .’’’’@! eiw"
 
# printing initial string
print ("initial string : ", ini_string)
 
result = re.sub('[\W_]+', '', ini_string)
 
# printing final string
print ("final string", result)

remove_symbols("L LIKED 50 Rolls Compatible with Dymo 30336 1)

# symb_re = re.compile(r"""[!"#$%&\'()*+,-./’:;<=>?@[\\\]^_`{|}~،؟…«“\":\"…”]""")
# symb_re2 = re.compile(r"""[\W_]+""")
symb_re3 = re.compile(r"""[^a-zA-Z0-9\s]+""")
def remove_symbols(text: str) -> str:
    return symb_re3.sub(repl="", string=text)

def replace_non_dates(date_i):
    if type(date_i) == str:
        if date_i[-1] =="m" or date_i[-1] =="s" or date_i[-1] =="h":
            date_i = date.today()
            return date_i
        elif "," not in date_i:
            if date_i.split()[-1].isnumeric():
                date_i = date_i + ", {}".format(date.today().year)
            else:
                date_i = date.today()
            return date_i
    return date_i

def convert_number(num_i):
    if type(num_i) == str:
        num_i = num_i.replace(',', '')
        if "K" in num_i:
            if "." in num_i:
                num_i = num_i.replace('.', '')[:-1]
                num_i = int(num_i) * 100
            else:
                num_i = int(num_i[:-1]) * 1000
        elif "M" in num_i:
            if "." in num_i:
                num_i = num_i.replace('.', '')[:-1]
                num_i = int(num_i) * 100000
            else:
                num_i = int(num_i[:-1]) * 1000000
        return num_i
    elif num_i == None:
        return -1
    else:
        return int(num_i)

## `1) Amazon Data`

# 1
amazon_product_data = pd.read_csv("../Datasets/product_amazon_data_pt1.csv")
amazon_product_data = amazon_product_data.rename(columns={"Unnamed: 0": "product_id_AMA"})
amazon_product_cols = amazon_product_data.columns.tolist() # columns list
# amazon_product_data = amazon_product_data.values # Numpy array

# CLEANING (should be added to the scraping process)
amazon_product_data["review_text"] = amazon_product_data["review_text"].astype("str")

amazon_product_data["product"] = list(map(lambda x: x.replace(":", ""), amazon_product_data['product']))
amazon_product_data["product"] = list(map(remove_symbols, amazon_product_data['product']))
amazon_product_data["amazon_product_name"] = list(map(lambda x: x.replace(":", ""), 
                                                          amazon_product_data['amazon_product_name']))
amazon_product_data["amazon_product_name"] = list(map(remove_symbols, 
                                                          amazon_product_data['amazon_product_name']))
amazon_product_data["review_text"] = list(map(lambda x: x.replace(":", ""), 
                                                          amazon_product_data['review_text']))
amazon_product_data["review_text"] = list(map(remove_symbols, 
                                                          amazon_product_data['review_text']))

amazon_product_data["rating"] = list(map(lambda x: x.split()[0], amazon_product_data['rating']))
amazon_product_data["rating"] = amazon_product_data["rating"].astype(float)

amazon_product_data["review_count"] = list(map(lambda x: x.split()[0].replace(",", ""), amazon_product_data['review_count']))
amazon_product_data["review_count"] = amazon_product_data["review_count"].astype(int)

amazon_product_data["price"] =  list(map(lambda x: x[1:] if type(x)==str else -1, 
                                         amazon_product_data['price']))
amazon_product_data["price"] = amazon_product_data["price"].astype(float)

amazon_product_data['review_date'] = list(map(replace_non_dates, amazon_product_data['review_date']))
amazon_product_data['review_date'] = pd.to_datetime(amazon_product_data['review_date'])


# create/use database
dbname = "amazon_product_reviews_db"
AMA_mongo_DB = get_database(db_name=dbname)
print("CREATE / USE database - '{}'".format(dbname))

# create/use collection
collname = "product_reviews"
mongo_coll = AMA_mongo_DB[collname]
print("CREATE / USE collection - '{}'".format(collname))
print()

# # empty collection
mongo_coll.delete_many({})

# insert data
insert_batch = 1000
batch_num = math.ceil(amazon_product_data.shape[0] / 1000)

for i in range(batch_num):
    data_insert = amazon_product_data[(i)*insert_batch: (i+1)*insert_batch]
    mongo_coll.insert_many(data_insert.to_dict(orient="records"))
    print("Inserted - {} documents".format(data_insert.shape[0]))

## `2) Twitter Data`

twitter_product_data = pd.read_csv("../Datasets/product_twitter_data_pt2_2.csv")
twitter_product_data = twitter_product_data.rename(columns={"Unnamed: 0": "tweet_id"})
twitter_product_cols = twitter_product_data.columns.tolist()

# CLEANING (should be added to the scraping process)
twitter_product_data["product"] = twitter_product_data["product"].astype("str")
twitter_product_data["tweet_poster"] = twitter_product_data["tweet_poster"].astype("str")
twitter_product_data["tweet_text"] = twitter_product_data["tweet_text"].astype("str")
twitter_product_data["product"] = list(map(lambda x: x.replace(":", ""), twitter_product_data['product']))
twitter_product_data["tweet_poster"] = list(map(lambda x: x.replace(":", ""), 
                                                     twitter_product_data['tweet_poster']))
twitter_product_data["tweet_poster"] = list(map(remove_symbols, 
                                                     twitter_product_data['tweet_poster']))
twitter_product_data["tweet_mention"] = list(map(lambda x: x.replace(":", ""), 
                                                     twitter_product_data['tweet_mention']))
twitter_product_data["tweet_mention"] = list(map(remove_symbols, 
                                                     twitter_product_data['tweet_mention']))
twitter_product_data["tweet_text"] = list(map(lambda x: x.replace(":", ""), 
                                                     twitter_product_data['tweet_text']))
twitter_product_data["tweet_text"] = list(map(remove_symbols, 
                                                     twitter_product_data['tweet_text']))

twitter_product_data["num_comments"] = list(map(convert_number, twitter_product_data['num_comments']))
twitter_product_data["num_comments"] =  twitter_product_data["num_comments"].astype(int)

twitter_product_data["num_retweets"] = list(map(convert_number, twitter_product_data['num_retweets']))
twitter_product_data["num_retweets"] =  twitter_product_data["num_retweets"].astype(int)

twitter_product_data["num_likes"] = list(map(convert_number, twitter_product_data['num_likes']))
twitter_product_data["num_likes"] =  twitter_product_data["num_likes"].astype(int)

twitter_product_data["num_views"] = list(map(convert_number, twitter_product_data['num_views']))
twitter_product_data["num_views"] =  twitter_product_data["num_views"].astype(int)

twitter_product_data['tweet_date'] = list(map(replace_non_dates, twitter_product_data['tweet_date']))
twitter_product_data['tweet_date'] = pd.to_datetime(twitter_product_data['tweet_date'])

# create/use database
dbname = "twitter_sentiment_db"
mongo_DB = get_database(db_name=dbname)
print("CREATE / USE database - '{}'".format(dbname))

# create/use collection
collname = "tweet_collection"
mongo_coll = mongo_DB[collname]
print("CREATE / USE collection - '{}'".format(collname))
print()

# # empty collection
mongo_coll.delete_many({})

# insert data
insert_batch = 1000
batch_num = math.ceil(twitter_product_data.shape[0] / 1000)

for i in range(batch_num):
    data_insert = twitter_product_data[(i)*insert_batch: (i+1)*insert_batch]
    mongo_coll.insert_many(data_insert.to_dict(orient="records"))
    print("Inserted - {} documents".format(data_insert.shape[0]))
