==> notebooks/10 - PostgreSQL Data.ipynb <==
import psycopg2
from psycopg2.extras import execute_values

import numpy as np
import pandas as pd

dojo_shopping_product_df = pd.read_csv("../Datasets/shopping_product_data.csv")
dojo_shopping_product_df = dojo_shopping_product_df.rename(columns={"Unnamed: 0": "product_id",
                                                                    "product_type": "category",
                                                                    "product_subtype": "sub_category"})
dojo_shopping_product_df = dojo_shopping_product_df[["product_id", "product_name", "product_code", "category",
                                                            "sub_category", "product_price"]]

dojo_shopping_product_cols = dojo_shopping_product_df.columns.tolist()
dojo_shopping_product_data = dojo_shopping_product_df.values

### `Database Connection`

user = "postgres"           # postgres is default postgresql superuser
password = "123m"     # please provide password for postgres user

conn = psycopg2.connect(
    host="localhost",
    database="ds_products_db",
    user=user,
    password=password)

cur = conn.cursor()


### `Create Table`

sql_create_table_shopping_products = \
"""
CREATE TABLE IF NOT EXISTS ShoppingProducts (
    product_id int PRIMARY KEY,
    product_name varchar(300) NOT NULL,
    product_code varchar(300) NOT NULL,
    category varchar(100),
    sub_category varchar(100),
    product_price float
);
"""

cur.execute(sql_create_table_shopping_products)

### `Data Insertion`

print("INSERT - Shopping Products")
execute_values(cur, 'INSERT INTO ShoppingProducts VALUES %s', dojo_shopping_product_data.tolist())

### `Close Connection`

print("DONE!")
cur.close()
conn.commit()
