==> notebooks/11 - trying avro data.ipynb <==
import avro.schema
from avro.datafile import DataFileReader, DataFileWriter
from avro.io import DatumReader, DatumWriter

schema = avro.schema.parse(open("Avro-Schemas/user.avsc", "rb").read())

writer = DataFileWriter(open("users.avro", "wb"), DatumWriter(), schema)
writer.append({"name": "Alyssa", "favorite_number": 256})
writer.append({"name": "Ben", "favorite_number": 7, "favorite_color": "red"})
writer.close()

reader = DataFileReader(open("users.avro", "rb"), DatumReader())
for user in reader:
    print(user)
# reader.close()


reader = DataFileReader(open("users.avro", "rb"), DatumReader())
list(reader)

reader = DataFileReader(open("sales_feed+0+0000000009+0000000009.avro", "rb"), DatumReader())
list(reader) 

list(reader)
